% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_setup.R
\name{initial_setup}
\alias{initial_setup}
\title{initial setup pre-mapping}
\usage{
initial_setup(postcodes = NULL, local.authority.area = "SW")
}
\arguments{
\item{postcodes}{a character vector containing postcodes to map}
}
\value{
a list
}
\description{
A helper function to convert postcodes to coordinates and PHE geographies
             and read in shape files
}
