% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_basic.R, R/map_choro.R
\name{map_basic}
\alias{map_basic}
\title{Produce a basic interactive map}
\usage{
map_basic(postcodes = NULL, minzoom = NULL, maxzoom = NULL,
  include.popup = TRUE)

map_basic(postcodes = NULL, minzoom = NULL, maxzoom = NULL,
  include.popup = TRUE)
}
\arguments{
\item{postcodes}{a character vector containing postcodes to map}

\item{minzoom}{a numeric defining minimum map zoom}

\item{maxzoom}{a numeric defining maximum map zoom}

\item{include.popup}{a logical indicating if postcodes should be included as a popup}

\item{postcodes}{a character vector containing postcodes to map}

\item{minzoom}{a numeric defining minimum map zoom}

\item{maxzoom}{a numeric defining maximum map zoom}

\item{include.popup}{a logical indicating if postcodes should be included as a popup}
}
\value{
a data.frame consisting of data copied to clipboard

a data.frame consisting of data copied to clipboard
}
\description{
A function to easily copy data into R from a spreadsheet.
             This should not be used within a reproducible work flow, instead
             it should only be used on an ad-hoc basis
             (note: I'm not aware of the original author of this function,
             it was taken directly from a stack overflow answer)

A function to easily copy data into R from a spreadsheet.
             This should not be used within a reproducible work flow, instead
             it should only be used on an ad-hoc basis
             (note: I'm not aware of the original author of this function,
             it was taken directly from a stack overflow answer)
}
\examples{
map_basic(c("tw12ju", "bs29rq"))

# alternatively copy a list of postcodes to your clipboard (i.e. highlight postcodes in spreadsheet and press ctrl + c) and use function with no arguments

map_basic()
# first copy data to clipboard (i.e. highlight data in spreadsheet and press ctrl + c)

# apply function to read data from clipboard into R

data = read_from_clipboard()
}
